Change link description to use choice instead of all.
Unbounded elements in all are fine in XSD 1.1 but not valid in XSD 1.0.
There aren't a ton of technologies out in the wild that leverage XSD 1.1
(e.g. code generators). This definition should be acceptable for
validation of XML documents but I'm not convinced it's a nice way to do
things re: code generation, as the code generated by XJC for example
contains a single List field containing the inertial and the
collisions/visuals.