Switch to a class for version parsing.
The previous attempt used a float to store the version number.
That has two problems:

1.  It runs into locale issues, and
2.  The XSD specifies it as a string

This alternative introduces a class to do parsing of the version
string.  Besides conforming to the XSD, and getting us away
from the locale issues, it also allows us to provide additional
constraints on the version (like not allowing negative numbers).
The UDFVersion class ends up living in the header file so we
can write tests against it.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>