Deprecate the APIs that we think are unused.
In an earlier commit, we changed from tinyxml -> tinyxml2
in the public API because we thought that there were no
users of these APIs.  Codify that here by marking these
APIs as deprecated; if a user comes along and says that
they are actually using it, we can undeprecated it.

Note that in order to avoid deprecations from within the
library, I had to add a bit of additional indirection here.
If we remove the APIs in the future, we can also remove this
indirection.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>